pyproject.toml
[tool.poetry]
name = "chatgpt-fastapi"
version = "0.1.0"
description = ""
authors = ["Labidahrom <labidahrom@yandex.ru>"]
readme = "README.md"
packages = [{include = "chatgpt_fastapi"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
sqlalchemy = {extras = ["async"], version = "^2.0.22"}
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
jinja2 = "^3.1.2"
passlib = "^1.7.4"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
httpx = "^0.25.1"
openai = "^1.2.3"
psutil = "^5.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
